+++ src/main/java/ru/kata/spring/boot_security/demo/service/RoleService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface RoleService {
     List<Role> getAllRoles();
 
-    Role getRole(int id);
+    Role getRole(Long id);
     List<Role> getRolesByIds(List<Integer> roleIds);
 }
