+++ src/main/java/ru/kata/spring/boot_security/demo/service/RoleServiceImpl.java
@@ -20,7 +20,7 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Role getRole(int id) {
+    public Role getRole(Long id) {
         return roleDao.getRole(id);
     }
